+++ src/tsipy/utils/plot.py
@@ -106,7 +106,7 @@ def plot_signals(
     **kwargs: Any,
 ) -> Tuple[Figure, Axes]:
     """Helper function for plotting signals."""
-    # pylint: disable=C0103, R0913, R0914
+    # pylint: disable=R0913, R0914
     fig, ax = plt.subplots()
     for signal_fourplet in signal_fourplets:
         x, y, label, kwargs_sig = signal_fourplet
@@ -143,7 +143,7 @@ def plot_signals_and_confidence(
     **kwargs: Any,
 ) -> Tuple[Figure, Axes]:
     """Helper function for plotting signal mean and confidence interval."""
-    # pylint: disable=C0103, R0913, R0914
+    # pylint: disable=R0913, R0914
 
     # Computes confidence interval width for Normal(0, 1)
     factor = norm.ppf(1 / 2 + confidence / 2)  # 0.95 % -> 1.959963984540054
@@ -188,7 +188,7 @@ def plot_signals_history(
     **kwargs: Any,
 ) -> Tuple[Figure, Axes]:
     """Helper function for plotting degradation correction history."""
-    # pylint: disable=C0103, R0913, R0914
+    # pylint: disable=R0913, R0914
     fig, axs = plt.subplots(nrows=n_rows, ncols=n_cols, figsize=fig_size)
 
     for i, signals in enumerate(signals_history):